+++ common/src/main/java/top/skyeyefast/mchjong/item/MahjongCatalog.java
@@ -22,6 +22,7 @@ public final class MahjongCatalog {
         entries.add(new ItemStack(MahjongContent.MAHJONG_DYE));
         entries.add(new ItemStack(MahjongContent.CREATIVE_MAHJONG_DYE));
         entries.add(new ItemStack(MahjongContent.RED_DORA_DYE));
+        entries.add(new ItemStack(MahjongContent.UNDO_DYE));
         MahjongComponents.DENOMINATIONS.stream().sorted().forEach(points -> {
             var stick = new ItemStack(MahjongContent.POINT_STICK);
             stick.set(MahjongComponents.POINTS, points);