+++ common/src/main/java/top/skyeyefast/mchjong/world/MahjongContent.java
@@ -36,10 +36,12 @@ public final class MahjongContent {
     public static final Item BOX_ITEM = new MahjongBoxItem(new Item.Properties().stacksTo(1).component(DataComponents.CONTAINER, ItemContainerContents.EMPTY));
     public static final Item MAHJONG_DYE = new Item(new Item.Properties().stacksTo(64));
     public static final Item RED_DORA_DYE = new Item(new Item.Properties().stacksTo(64));
+    public static final Item UNDO_DYE = new Item(new Item.Properties().stacksTo(64));
     public static final Item CREATIVE_MAHJONG_DYE = new Item(new Item.Properties().stacksTo(1).rarity(net.minecraft.world.item.Rarity.EPIC));
     public static final java.util.Map<String, Item> SUPPLIES = java.util.Map.of(
         "table_cloth", CLOTH_ITEM, "mahjong_tile", TILE_ITEM, "point_stick", POINT_STICK, "mahjong_box", BOX_ITEM,
-        "mahjong_dye", MAHJONG_DYE, "creative_mahjong_dye", CREATIVE_MAHJONG_DYE, "red_dora_dye", RED_DORA_DYE);
+        "mahjong_dye", MAHJONG_DYE, "creative_mahjong_dye", CREATIVE_MAHJONG_DYE, "red_dora_dye", RED_DORA_DYE,
+        "undo_dye", UNDO_DYE);
     public static BlockEntityType<MahjongTableBlockEntity> TABLE_ENTITY;
     public static BlockEntityType<FurnitureBlockEntity> STOOL_ENTITY;
     public static EntityType<SeatEntity> SEAT_ENTITY;