+++ neoforge/src/main/java/top/skyeyefast/mchjong/neo/MchjongNeoForge.java
@@ -53,6 +53,7 @@ public final class MchjongNeoForge {
         items.register("point_stick", () -> MahjongContent.POINT_STICK);
         items.register("mahjong_dye", () -> MahjongContent.MAHJONG_DYE);
         items.register("red_dora_dye", () -> MahjongContent.RED_DORA_DYE);
+        items.register("undo_dye", () -> MahjongContent.UNDO_DYE);
         items.register("creative_mahjong_dye", () -> MahjongContent.CREATIVE_MAHJONG_DYE);
         items.register("mahjong_stool", () -> MahjongContent.STOOL_ITEM);
         items.register(bus);